refactor: moved lib folder from src to app
