+++ src/app/lib/assets.ts
