+++ src/app/lib/compare.ts
