+++ src/app/services/api/chart.service.ts
@@ -14,7 +14,7 @@ import {
 	MutateChartModel,
 } from "@/models/chart.model";
 
-import { apiUrl } from "../../../lib/api";
+import { apiUrl } from "../../lib/api";
 
 @Injectable({
 	providedIn: "root",
