+++ src/app/services/api/known-issue.service.ts
@@ -11,7 +11,7 @@ import type {
 	KnownIssueModel,
 } from "@/models/known-issue.model";
 
-import { apiUrl } from "../../../lib/api";
+import { apiUrl } from "../../lib/api";
 
 @Injectable({
 	providedIn: "root",
