+++ src/app/services/api/version.service.ts
@@ -8,7 +8,7 @@ import { CacheService } from "../cache.service";
 // Models
 import { CreateVersionModel, VersionModel } from "@/models/version.model";
 
-import { apiUrl } from "../../../lib/api";
+import { apiUrl } from "../../lib/api";
 
 @Injectable({
 	providedIn: "root",
