+++ tsconfig.json
@@ -19,15 +19,9 @@
 		"module": "ES2022",
 		"baseUrl": "src",
 		"paths": {
-			"@/*": [
-				"app/*"
-			],
-			"@/lib/*": [
-				"lib/*"
-			],
-			"@/assets/*": [
-				"assets/*"
-			]
+			"@/*": ["app/*"],
+			"@/lib/*": ["app/lib/*"],
+			"@/assets/*": ["assets/*"]
 		}
 	},
 	"angularCompilerOptions": {
